restructured context api hooks
