+++ pages/tokens/add.tsx
@@ -14,14 +14,13 @@ import {
   USER_CANCELED_TX,
 } from "../../lib/errors";
 import { TokenInfo } from "../../lib/types";
-import { useMessageAlert } from "../../lib/hooks/message-alert";
+import { useMessageAlert } from "../../lib/hooks/context/message-alert";
 import { useSigner } from "../../lib/hooks/useSigner";
-import { useStoredTokens } from "../../lib/hooks/tokens";
+import { useStoredTokens } from "../../lib/hooks/context/tokens";
 import { useScrollTop } from "../../lib/hooks/useScrollTop";
 import { EventType, trackEvent } from "../../lib/analytics";
 import { FORTY_DIGITS_HEX } from "../../lib/constants/regex";
 import { abbreviatedTokenAmount, shortAddress } from "../../lib/utils";
-import { ETHERSCAN_ADDRESS_PREFIX } from "../../lib/constants/url";
 
 import { Spinner } from "../../components/spinner";
 import SectionTitle from "../../components/sectionTitle";
@@ -179,11 +178,7 @@ const TokenContainer = ({ symbol, name, totalSupplyFormatted, address }) => (
       <Info>
         <TokenAttributeTitle>Token address</TokenAttributeTitle>
         {/* TODO insert copy icon (and use the same links as in header?) [VR 30-06-2021] */}
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href={`${ETHERSCAN_ADDRESS_PREFIX}/${address}`}
-        >
+        <a target="_blank" rel="noopener noreferrer" href={`https://etherscan.io/${address}`}>
           <Address>{shortAddress(address)}</Address>
         </a>
       </Info>
